refactor: 完成main.cpp模块化解耦
- 创建AppConfig.h集中管理配置常量
- 创建CommandProcessor.h/cpp封装MQTT命令处理
- 创建RfidFrameParser.h/cpp封装RFID帧解析
- 创建SystemManager.h/cpp统一管理系统初始化和主循环
- 简化main.cpp为15行入口文件
- 修复编译错误：NetworkStatus/NetworkType声明、回调签名匹配、heartbeatInterval类型

+++ include/ConfigManager.h
@@ -19,7 +19,7 @@ typedef struct {
     char subTopic[64];
     
     uint16_t alarmDuration;
-    uint16_t heartbeatInterval;
+    uint32_t heartbeatInterval;
     uint8_t bleEnabled;
 } SystemConfig;
